Code Refactor
Instead of adding the scrape function in main.py file, created a separate module for it following the code convention.
