+++ src/scrape_up/BBCNews/__init__.py
@@ -1,0 +1,3 @@
+from bbc_news import BBCNewsScraper
+
+__all__=["BBCNewsScraper"]
